// api/src/types/news.ts
export interface IinitialState {
  news: any[]
  loading: boolean
  error: null | string
}

export enum NewsActionTypes {
  FETCH_NEWS = 'FETCH_NEWS',
  FETCH_NEWS_SUCCESS = 'FETCH_NEWS_SUCCESS',
  FETCH_NEWS_ERROR = 'FETCH_NEWS_ERROR',
}

interface FetchNewsAction {
  type: NewsActionTypes.FETCH_NEWS
}

interface FetchNewsSuccessAction {
  type: NewsActionTypes.FETCH_NEWS_SUCCESS
  payload: any[]
}
interface FetchNewsErrorAction {
  type: NewsActionTypes.FETCH_NEWS_ERROR
  payload: string
}

export type NewsAction =
  | FetchNewsAction
  | FetchNewsErrorAction
  | FetchNewsSuccessAction